split table to header and body
