+++ packages/sogrim-app/src/components/Pages/SemestersPage/SemesterTable/SemesterTable.tsx
@@ -2,72 +2,17 @@ import * as React from 'react';
 import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
-import TableSortLabel from '@mui/material/TableSortLabel';
-import { visuallyHidden } from '@mui/utils';
 import { observer } from 'mobx-react-lite';
 import { getComparator, Order, stableSort } from './SemesterTableUtils';
-import { RowData, headCells } from './SemesterTabsConsts';
+import { RowData } from './SemesterTabsConsts';
 import { SemesterTableRow } from './SemesterTableRow';
 import { Paper } from '@mui/material';
 import useUserState from '../../../../hooks/apiHooks/useUserState';
 import { useAuth } from "../../../../hooks/useAuth";
 import { useStore } from "../../../../hooks/useStore";
 import useUpdateUserState from '../../../../hooks/apiHooks/useUpdateUserState';
-
-interface EnhancedTableProps {
-  onRequestSort: (event: React.MouseEvent<unknown>, property: keyof RowData) => void;
-  order: Order;
-  orderBy: string;
-}
-
-const EnhancedTableHead: React.FC<EnhancedTableProps> = ({
-    order, orderBy, onRequestSort
-}) => {
-  
-  const createSortHandler =
-    (property: keyof RowData) => (event: React.MouseEvent<unknown>) => {
-      onRequestSort(event, property);
-    };
-
-  return (
-    <TableHead>
-      <TableRow>        
-        {headCells.map((headCell) => (
-          <TableCell
-            align='center'
-            key={headCell.id}            
-            padding={headCell.disablePadding ? 'none' : 'normal'}
-            sortDirection={orderBy === headCell.id ? order : false}
-          >
-            <TableSortLabel
-              active={orderBy === headCell.id}
-              direction={orderBy === headCell.id ? order : 'asc'}
-              onClick={createSortHandler(headCell.id)}
-            >
-              {headCell.label}
-              {orderBy === headCell.id ? (
-                <Box component="span" sx={visuallyHidden}>
-                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                </Box>
-              ) : null}
-            </TableSortLabel>
-          </TableCell>
-        ))}
-        <TableCell
-            align='center'
-            key={'header-actions'}                                    
-          >
-              פעולות
-          </TableCell>
-      </TableRow>
-    </TableHead>
-  );
-}
-
+import { SemesterTableHeader } from './SemesterTableHeader';
 export interface SemesterTableProps {
     rows: RowData[],
     semester: string;
@@ -114,7 +59,7 @@ const SemesterTableComp: React.FC<SemesterTableProps> = ({
             aria-labelledby="tableTitle"
             size={'small'}
           >
-            <EnhancedTableHead
+            <SemesterTableHeader
               order={order}
               orderBy={orderBy}
               onRequestSort={handleRequestSort}
@@ -124,7 +69,7 @@ const SemesterTableComp: React.FC<SemesterTableProps> = ({
                 .map((row, index) => {                  
                   const labelId = `enhanced-table-checkbox-${index}`;
                   return (
-                      < SemesterTableRow row={row} labelId={labelId} handleSave={handleSave} key={index} />                   
+                      <SemesterTableRow row={row} labelId={labelId} handleSave={handleSave} key={index} />                   
                   );
                 })}              
             </TableBody>
